Specify a single host to set up or run tests
Using 'all' instead of '127.0.0.1' gives you access to group_vars, at
cost of running the play multiple times. Choosing a single host is
confusing, as the playbook is actually run on localhost. It's also less
portable, as it's based on this particular project's inventory. However,
this currently seems the best compromise.