Deploy catalyst with Capistrano
This changes the `catalyst_install` playbook
to use Capistrano for deploying the rails app.

The basic strategy for the deployment is cloning
the git repo to the home folder of the user who
is deploying. The user needs to be in the `catalyst`
group.

In that folder a localhost cap deploy is ran
that deploys to `/opt/catalyst`. That directory needs
read/write permissions for the `catalyst` group.

The apache virtualhost config needs to point to
`/opt/catalyst/current/public`. Capistrano creates the
`current` symlink to point to the latest release. It
keeps 5 releases that can be rolled back to.

This commit also tightens up the install for `yarn` and
`webpack`, which need to be installed before `rails assets:precompile`
is ran as part of the deploy process. Those installs
use conditionals so that they will only be installed if they
are not already present.

This should only be merged when the `capify` branch
for `catalyst-blackight` has been merged.