Ruby266 (#58)
* ruby-2.5.5

* ruby 2.6.6

* bundler 2.1.4

* jruby 9.2.13 update

* Update the script to pull from the jhu-library-applications/catalyst-blacklight repo.
Changed the default branch to be main in prod and stage and the develop branch in test and dev. This means that you don't have to add --extra-vars "app_branch={branch}" to deploy the main branch to catalyst.library.jhu.edu and catalyst-stage.library.jhu.edu or the develop branch to catalyst-test.library.jhu.edu

* Update ruby version to 2.6.6

* Added .idea directory to .gitignore

* Add bundler install

This explicit install is needed when bundlers lock files
are generated with the newer version.

* LAG-3847

update BDtest URL

* Webpack/yarn installation and bundler updates for version 2 (#57)

* Rollbar setup (#60)

This adds the rollbar access token to the prod
vault and uses it in the dotenv file.

* Update version comparison syntax (#61)

The syntax for doing a conditional based
on a version number has changed in ansible
2.9.x and this line will fail now.

This changes it to use the newer syntax.

See https://github.com/ansible/ansible/issues/64174

* Traject ruby version updates (#62)

This requires https://github.com/jhu-sheridan-libraries/catalyst-traject/tree/bundler-update
to be merged so that it uses the correct version of `bundler`

This installs the newer version of jruby we are using
with traject when running the install_traject playbook.

After running this you should be able to run
```
RAILS_ENV=production bundle exec traject -c conf/horizon_source.rb -c conf/horizon_index.rb -c conf/solr_connect.rb -s horizon.only_bib=7364665
```
without errors.

On `catsolrmaster-test` solr wasn't running. It can be started like this:

`sudo systemctl start solr`

* Use the bundler module to install gems (#63)

This changes the bundle installation process
to use the ansible module. This raises
deprecation warnings because bundler v2 has
changed the way they are doing configuration.

This should keep working until those deprecations
raise errors or until the ansible module is updated.

Co-authored-by: Jon Fackrell <jon.fackrell@jhu.edu>
Co-authored-by: Jamie Little <jamie@jamielittle.org>
Co-authored-by: fackrellj <75277580+fackrellj@users.noreply.github.com>
Co-authored-by: Derek Belrose <dbelrose@jhu.edu>