Use systemd to manage solr service
This changes the way the solr service is managed
on the VM so that it uses `systemd`.

This way we can have the service automatically
restart if it goes down. Over the past few weeks
we have had the `oom_memory_killer` script for
solr run and then solr isn't restarted.